service/requirements.txt
lxml>=4.6.1
absl-py>=0.10.0
matplotlib>=3.0.3
numpy>=1.19.4
Pillow>=8.0.0
PyYAML>=5.1
six>=1.15.0
tensorflow>=2.4.0
tensorflow-addons>=0.12
tensorflow-hub>=0.11
neural-structured-learning>=1.3.1
tensorflow-model-optimization>=0.5
Cython>=0.29.13
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
opencv-contrib-python>=4.5.1.48
flask>=1.1.1
flask-cors>=3.0.10